# feat(query): Improve query performance