… query/iterator.gen.go