As a follow on to #17596, performance of both sort and unsorted merge operations can be improved significantly by removing allocations when comparing tags.

This PR makes the following improvements:

* Adds a method, `CompareKeys`, to the `Tags` struct in `points.go`. `CompareKeys` performs a lexicographical comparison of the two `Tags` using the specified `keys` without performing any allocations. The comparison of the values is in the order determined by `keys`. This is important in maintaining compatibility with the replaced `Tags#ID` sort behavior.

* Updates the `Less` methods of `<type>MergeHeap` and `<type>SortedMergeHeap`, replacing multiple calls to the costly `Tags#Subset` API with the `CompareKeys` API.

* In order to keep the same sort behavior, `<type>MergeIterator` and `<type>SortedMergeIterator` were updated to store a cached, sorted list of the `IteratorOptions#Dimensions` slice, which is passed to `CompareKeys`. This ensures the Tags sort in lexicographical order based on the lexicographical sort order of the Dimensions slice. Doing this retains the same sort behavior as the `Tags#ID` API, previously used for comparisons.

`benchstat` results:

```
name                    old time/op    new time/op    delta
SortedMergeIterator-16    32.4ms ± 2%     5.2ms ± 3%  -83.81%  (p=0.000 n=10+10)

name                    old alloc/op   new alloc/op   delta
SortedMergeIterator-16    36.5MB ± 0%     5.8MB ± 0%  -84.20%  (p=0.000 n=10+9)

name                    old allocs/op  new allocs/op  delta
SortedMergeIterator-16      420k ± 0%       60k ± 0%  -85.71%  (p=0.000 n=9+10)
```